// backend/src/modules/bookings/booking.enums.ts
export enum BookingStatus {
    PENDING = "pending",

    CONFIRMED = "confirmed",

    COMPLETED = "completed",

    CANCELLED = "cancelled",

    FAILED = "failed",

}
export enum PaymentMethod {
    CASH = "cash",

    PAYHERE = "payhere",
}
export enum PaymentStatus {
    PENDING = "pending",

    UNPAID = "unpaid",

    PAID = "paid",

    FAILED = "failed",

    REFUNDED = "refunded",
}